requirements.txt
pynetdicom
pydicom
python-dotenv
requests